⚡️ feat: Add @nestjs/cache-manager for caching
This commit adds the "@nestjs/cache-manager" package as a new dependency in the project. It also imports the CacheModule from the package in the "translate.module.ts" file and registers it in the module. This change enables caching functionality in the application, which can improve performance by storing frequently accessed data in memory.